Allow to create images as well as chaging them
Why:

* There are situations when it is useful to be able to create images,
  which is supported by imagemagick. As far as I could tell, it is only
  supported by the `convert` command, not `mogrify`.

This change addresses the need by:

* Separating the concepts of saving and creating, being that saving will
  use mogrify and requires an actual image file to be present beforehand
  whereas creating allows you to create a virtual canvas and then save it
  to the file.
* Because of this separation, the name of the path for creating won't
  work with the `-write` flag.
* The imagemagick syntax for creating a canvas is part of what is called
  image operators, and is something like `xc:white`, which means the
  assumption that all argument have a name prefixed with `-` and a value
  does not hold true. I think image operators the only options that do not
  follow that syntax, so I added a special case for it.

The name of the project (Mogrify) suggest it is a wrapper only to that
command, but the description seems to be more broad, as it says it's a
wrapper for imagemagick. Because of that, I am unsure if this is
something that is interesting and/or fits this project, but I wanted to
have this discussion before thiking of forking it.